// darkhorse-dotnet/toolkit/packages/gantt/src/types.ts
export type GanttToolbarAction = 
  | 'Add' 
  | 'Edit' 
  | 'Delete' 
  | 'ZoomIn' 
  | 'ZoomOut' 
  | 'ZoomToFit' 
  | 'Undo' 
  | 'Redo' 
  | 'ExpandAll' 
  | 'CollapseAll';